add missing functions for abb_egm_rws_managers library
abb_egm_rws_managers is a library that allows to use EGM and RWS with
ROS by providing a higher-level management to both EGM and RWS, as well
as a structured description of an ABB robot controller.

Add some missing functions to V2_0 (some were already present, just
needed to be declared as public methods) for V2_0 to be usable witb
abb_egm_rsw_managers.